--- PlayerToken.java
@@ -1,2 +1,0 @@
-package game.logic;
-public enum PlayerToken { NOPLAYER, PLAYER1, PLAYER2 };
